requirements_test.txt
# =============================================================================
# OMNI WRITER - REQUIREMENTS TEST ATUALIZADO
# =============================================================================
# Data: 2025-01-27
# Tracing ID: REQ_TEST_UPDATE_20250127_001
# Análise: Baseada em análise completa do sistema de testes
# =============================================================================

# =============================================================================
# FRAMEWORK DE TESTES PRINCIPAL
# =============================================================================
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-xdist>=3.0.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0
pytest-benchmark>=4.0.0
pytest-profiling>=1.7.0

# =============================================================================
# TESTES DE CARGA E PERFORMANCE
# =============================================================================
locust>=2.15.0
pytest-benchmark>=4.0.0

# =============================================================================
# TESTES VISUAIS E E2E
# =============================================================================
selenium>=4.0.0
webdriver-manager>=3.8.0
percy>=2.0.0
axe-selenium-python>=4.8.0

# =============================================================================
# TESTES DE MUTAÇÃO
# =============================================================================
mutmut==2.4.4

# =============================================================================
# TESTES DE SEGURANÇA
# =============================================================================
bandit>=1.7.0
safety>=2.3.0

# =============================================================================
# COBERTURA DE CÓDIGO
# =============================================================================
coverage>=7.3.0
pytest-cov>=4.1.0

# =============================================================================
# MOCKS E FIXTURES
# =============================================================================
responses>=0.23.0
freezegun>=1.2.0
factory-boy>=3.3.0

# =============================================================================
# TESTES DE INTEGRAÇÃO
# =============================================================================
requests-mock>=1.11.0
httpretty>=1.1.4

# =============================================================================
# TESTES DE BANCO DE DADOS
# =============================================================================
pytest-postgresql>=4.1.0
pytest-redis>=2.0.0

# =============================================================================
# TESTES DE API
# =============================================================================
pytest-flask>=1.3.0
pytest-fastapi>=0.1.0

# =============================================================================
# UTILITÁRIOS PARA TESTES
# =============================================================================
faker>=19.0.0
pytest-randomly>=3.15.0
pytest-repeat>=0.9.1

# =============================================================================
# NOTAS DE INSTALAÇÃO
# =============================================================================
# Para instalação completa de testes:
# pip install -r requirements_test.txt

# Para testes específicos:
# pip install pytest pytest-cov  # Testes básicos
# pip install locust  # Testes de carga
# pip install selenium webdriver-manager  # Testes visuais 